pyproject.toml
[tool.poetry]
name = "ysmirnov-dz-27-avito"
version = "0.1.0"
description = ""
authors = ["yasmirnov <iaroslav.irnovsm@gmail.com>"]
readme = "README.md"
packages = [{include = "ysmirnov_dz_27_avito"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
psycopg2 = "^2.9.6"
psycopg = "^3.1.9"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
